+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Use `property` instead of `name` attribute.
+
 ## [1.0.0] - 2019-06-13
 
 ### Added